pretraining/data_processing/requirements.txt
datasets
numpy
tokenizers
huggingface_hub
pyyaml
ruff
tqdm
mosaicml-streaming==0.8.1
